moved the language class
