+++ src/com/gmail/xfrednet/xfutils/util/language/Language.java
@@ -1,4 +1,4 @@
-package com.gmail.xfrednet.xfutils.util;
+package com.gmail.xfrednet.xfutils.util.language;
 
 import java.io.IOException;
 import java.util.Locale;
@@ -7,6 +7,7 @@ import java.util.ResourceBundle;
 import java.util.Set;
 
 import com.gmail.xfrednet.xfutils.Main;
+import com.gmail.xfrednet.xfutils.util.Settings;
 
 import javax.swing.*;
